Fixed E-sarsa and intialization
Fixed two issues that were causing E-sarsa and q learning to perform badly. Set initialiazation to -100 in each cell rather than 0 to reduce unnecessary exploration. Fixed the code for estimating the expectation of the next action, which was off-by-one on the number of actions and subtracted an unnecessary term so total p < 1.

Also fixed the fourth graph, which displayed the third graph earlier.